+++ ArcPicker/ArcPickerBehaviorDecorator.js
